Read the real audio format instead of asking for one, add conformance vectors
`AVCaptureAudioDataOutput.audioSettings` is macOS-only, so on iOS the
session decides the capture format. AudioStreamer now reads the stream
description off each buffer and converts to interleaved Int16 only when it
has to, refusing layouts it cannot test rather than guessing at them.

protocol/vectors/v1.json is generated by tools/ProtocolVectors and read by
BOTH test suites. Each implementation can be internally consistent and
still disagree with the other -- a different byte order, a different JSON
escape, a different HKDF label -- and nothing would catch it until two real
devices failed to pair. The vectors pin down framing, the key schedule,
sealed AEAD records, canonical handshake JSON including a non-ASCII device
name, media headers, and the control payloads.

48 tests pass on the Windows side.

+++ ios/iCam/Camera/CameraState.swift
@@ -44,7 +44,8 @@ enum CaptureTarget: String, Codable, Sendable, CaseIterable {
 enum WhiteBalancePreset: String, Codable, Sendable, CaseIterable {
     case auto, daylight, cloudy, shade, tungsten, fluorescent, warm, custom
 
-    var temperature: Float? {
+    /// Kelvin, or `nil` for the two presets that are not a fixed temperature.
+    var temperature: Double? {
         switch self {
         case .auto, .custom: return nil
         case .daylight:      return 5600
